Another failed attempt
So you can't enable DMA during hblank, so that idea's bust. However, it should still be possible to upload a byte or two to VRAM by using various HDMA channels to set the VRAM addr regs and write the data. In that case, we should generate all of the values to be written into HDMA channels during VBLANK and enable HDMA, then let the hw do the rest. A bit like queueing the updates we would've made to the characters, and encoding them in a bit of a strange way. A lot simpler than what I was thinking before, and in a way less fun.. but should actually work and play nicely with the hw. :)

Anyways, gonna call it quits for tonight and pick this up next week on the stream. This was fun! :)