Foiled again :)
Tried this time to use HDMA to update the VRAM graphics each line, but after a number of failed attempts and digging in emulator code, I found this isn't the right approach for this platform, as you can't actually write to VRAM outside of VBLANK, regardless of whether you're doing it from DMA, HDMA, CPU, etc. You just don't have VRAM access.

Interestingly enough I dug a tiny bit into the original intro's code that inspired me to do this, and it's actually displayed sideways using mode7, which is funny. I didn't dig into it too much though, as that's pretty out-of-scope for what I wanted to do with this on stream, so calling it quits for now and just pushing what I have :)